Entity 클래스에서 사용하는 enum 들을 디렉토리에 따로 두지않고 해당 클래스에 정의해두었다. 이유는 EnumClass 는 어디에 정의할까 text 파일에 작성
