+++ text/0. 패키지 구조
@@ -14,7 +14,7 @@
         - exception
         - repository
         - service
-        - type
+        - type (모든 클래스에서 공통적으로 사용할 것만 넣을예정, 특정 Entity 클래스와 관련있는것은 Entity 클래스 내부에 enum 작성)
 -test
     - java
         - controller
@@ -71,6 +71,8 @@
     뱉어낼지를 에러를 핸들링 해줄 것들이 들어갈 패키지 이다.
 
     - type 패키지
+        - (원래는 Enum 을 이곳에서 사용했으나 enum 들을 클래스와 관련있는 Entity 로 옮겨 직접적으로 관계가 있음을 코드상으로 표현해주었다.)
+        - (모든 클래스에서 공통적으로 사용할 것만 넣을예정, 특정 Entity 클래스와 관련있는것은 Entity 클래스 내부에 enum 작성)
     Enum 클래스들을 모아둘 공간이다. Enum 은 관련이 있는 상수들의 집합이다.
     자바에서는 final 로 String 과 같은 문자열이나 숫자들을 나타내는 기본 자료형의 값을
     고정할 수 있는데 이를 상수(변하지 않는 값)라고한다.
