+++ text/EnumClass는 어디에 정의할까
@@ -1,0 +1,19 @@
+* Enum 은 어디에 정의할까?
+
+    - 나는 강의를 보고 Enum 은 Type 이라는 디렉토리를 만들어
+      그 디렉토리 안에서 작성하여 사용하였다.
+      왜냐하면 DB, ORM 관련 클래스이기 때문에 이곳에 다른 무언가가 들어가는 바람직하지 못하다는
+      의견을 들었기 때문이었다.
+
+      그러나 기억보단 기록을(향로)님의 블로그를 보고 어떻게 사용하면 좋을까? 를 다시
+      생각하게 되었다.
+
+      Method enum 을 Payment 클래스 내부에 작성하였습니다.
+      이렇게 하면 enum 이 Payment 클래스와 직접적으로 관계가 있음을 코드상으로 표현할 수 있을 뿐더러,
+      다른 클래스에서 사용할때도 Payment.Method 가 되어 이름의 중복없이 깔끔하게 사용할 수 있습니다.
+                                                                                - By 기억보단 기록을(향로)
+
+      클래스와 직접적으로 관계가 있음을 코드상으로 표현할 수 있을 뿐만 아니라
+      좀 더 쉽게 타입에 무엇이 있는지 어떻게 정의되어 있는지 파악할 수 있어서 편하였고,
+      다른 사람이 인수인계를 받게되어도 해당 클래스와 관련있는 enum 들을 좀 더 빠르게 파악할 수 있는 장점이 생겼다.
+
